refactor(index): merge `src/rlb.js` and `src/index.js`
Closes #294.
